Split to two classes and update explanation.
